Uploaded the first version I had prepared and its source.
This verison is moderately useful. May accidentally delete all your files and has a little UI bug that I am aware of, namely displaying an empty path field as error rather than a lack of parameters. Cosmetic problem.

The code is very ugly because I only intend to continue work on it if there is any interest in it.